+++ src/components/Icons/iconChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      "Added the `fileDownloadAlt` icon.",
+    ],
+  },
+  {
     date: "2026-06-09",
     version: "4.2.0",
     type: "Update",